+++ src/test/java/com/puppycrawl/tools/checkstyle/checks/coding/StringLiteralEqualityCheckTest.java
@@ -35,7 +35,9 @@ public class StringLiteralEqualityCheckTest
     @Override
     protected String getPath(String filename) throws IOException {
         return super.getPath("checks" + File.separator
-                + "coding" + File.separator + filename);
+                + "coding" + File.separator
+                + "stringliteralequality" + File.separator
+                + filename);
     }
 
     @Test
